Add avg lux calculation

+++ master/threshold.h
@@ -8,12 +8,12 @@ class Threshold
   int low, high;
   bool toggle = false;
 
-  unsigned int wait;
+  unsigned long wait;
   unsigned long prev_time;
   bool waiting = false;
 
 public:
-  Threshold(int low, int high, int wait) : low(low), high(high), wait(wait) {}
+  Threshold(int low, int high, unsigned long wait) : low(low), high(high), wait(wait) {}
   void loop(int lux);
   bool toggled();
 };
